+++ smartcontract/programs/doublezero-serviceability/tests/device_test.rs
@@ -115,7 +115,7 @@ async fn test_device() {
             name: "Los Angeles".to_string(),
             lat: 1.234,
             lng: 4.567,
-            bgp_community: 0,
+            bgp_community: 10000,
         }),
         vec![
             AccountMeta::new(exchange_pubkey, false),
@@ -636,7 +636,7 @@ async fn setup_program_with_location_and_exchange(
             name: "Los Angeles".to_string(),
             lat: 1.234,
             lng: 4.567,
-            bgp_community: 0,
+            bgp_community: 10000,
         }),
         vec![
             AccountMeta::new(exchange_pubkey, false),