+++ smartcontract/programs/doublezero-serviceability/tests/link_dzx_test.rs
@@ -122,7 +122,7 @@ async fn test_dzx_link() {
             name: "Los Angeles".to_string(),
             lat: 1.234,
             lng: 4.567,
-            bgp_community: 0,
+            bgp_community: 10000,
         }),
         vec![
             AccountMeta::new(exchange_pubkey, false),