+++ smartcontract/programs/doublezero-telemetry/tests/test_helpers.rs
@@ -101,7 +101,7 @@ impl ExchangeCreateArgsExt for ExchangeCreateArgs {
             name: "".to_string(),
             lat: 0.0,
             lng: 0.0,
-            bgp_community: 0,
+            bgp_community: 10000,
         }
     }
 }
@@ -296,7 +296,7 @@ impl LedgerHelper {
             .create_exchange(ExchangeCreateArgs {
                 code: "EX1".to_string(),
                 name: "Test Exchange".to_string(),
-                bgp_community: 1,
+                bgp_community: 10001,
                 ..ExchangeCreateArgs::default()
             })
             .await?;
@@ -395,7 +395,7 @@ impl LedgerHelper {
             .create_exchange(ExchangeCreateArgs {
                 code: "EX1".to_string(),
                 name: "Test Exchange1".to_string(),
-                bgp_community: 1,
+                bgp_community: 10001,
                 ..ExchangeCreateArgs::default()
             })
             .await?;
@@ -405,7 +405,7 @@ impl LedgerHelper {
             .create_exchange(ExchangeCreateArgs {
                 code: "EX2".to_string(),
                 name: "Test Exchange2".to_string(),
-                bgp_community: 2,
+                bgp_community: 10002,
                 ..ExchangeCreateArgs::default()
             })
             .await?;